+++ src/models/TaxRate.php
@@ -7,12 +7,11 @@
 
 namespace craft\commerce\models;
 
-use Craft;
 use craft\commerce\base\Model;
+use craft\commerce\helpers\Localization;
 use craft\commerce\Plugin;
 use craft\commerce\records\TaxRate as TaxRateRecord;
 use craft\helpers\UrlHelper;
-use craft\i18n\Locale;
 use DateTime;
 
 /**
@@ -144,10 +143,7 @@ class TaxRate extends Model
      */
     public function getRateAsPercent(): string
     {
-        $percentSign = Craft::$app->getLocale()->getNumberSymbol(Locale::SYMBOL_PERCENT);
-        $rate = $this->rate * 100;
-
-        return Craft::$app->getFormatter()->asDecimal($rate) . $percentSign;
+        return Localization::formatAsPercentage($this->rate);
     }
 
     /**